requirements.txt
requests
python-dotenv
bcrypt==4.2.0

fastapi
uvicorn[standard]

python-multipart
